Added Procedures for Ease of adding Users and Superusers
Added Procedures to addAdmin and addSuperuser. These procedures will automatically create a row for them in the Users table before adding to the appropriate table.

If necessary, can do same for Events.

Just note that adding to child tables requires adding to parent tables first.

Example: To add to RSO Events, must first add to Events with the same event_id
AND
to add to Event, must first add to locations with same location id.